C0003_MCP_SSE/requirements.txt
pydantic
typing
psycopg2-binary
mcp